requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
transformers
rembg
httpx
onnxruntime
kornia
timm